requirements.txt
feedparser
yagmail
pyyaml
python-dateutil
translate